improved modularity of processing (images and JSONifying), gemini, and gpt4o
